remove the non null assertions

+++ src/index.ts
@@ -14,7 +14,7 @@ const pool = new Pool({
   connectionTimeoutMillis: 1000,
 });
 
-const streamClient = createClient(EvmStream, process.env.APIBARA_URL!);
+const streamClient = createClient(EvmStream, process.env.APIBARA_URL);
 
 function msToHumanShort(ms: number): string {
   const units = [
@@ -166,31 +166,32 @@ const refreshAnalyticalTables = throttle(
         const isHead = message.data.production === "live";
 
         for (const block of message.data.data) {
-          const blockNumber = Number(block!.header!.blockNumber);
+          if (!block) continue;
+          const blockNumber = Number(block.header.blockNumber);
           deletedCount += await dao.deleteOldBlockNumbers(blockNumber);
 
-          const blockTime = block!.header!.timestamp!;
+          const blockTime = block.header.timestamp;
 
           await dao.insertBlock({
-            hash: BigInt(block!.header!.blockHash ?? 0),
-            number: block!.header!.blockNumber,
+            hash: BigInt(block.header.blockHash ?? 0),
+            number: block.header.blockNumber,
             time: blockTime,
           });
 
-          for (const event of block!.logs) {
+          for (const event of block.logs) {
             const eventKey: EventKey = {
               blockNumber,
-              transactionIndex: event.transactionIndex!,
-              eventIndex: event.logIndexInTransaction!,
-              emitter: BigInt(event.address!),
-              transactionHash: BigInt(event.transactionHash!),
+              transactionIndex: event.transactionIndex,
+              eventIndex: event.logIndexInTransaction,
+              emitter: BigInt(event.address),
+              transactionHash: BigInt(event.transactionHash),
             };
 
             // process each event sequentially through all the event processors in parallel
             // assumption is that none of the event processors operate on the same events, i.e. have the same filters
             // this assumption could be validated at runtime
             await Promise.all(
-              event.filterIds!.map(async (matchingFilterId) => {
+              event.filterIds.map(async (matchingFilterId) => {
                 eventsProcessed++;
 
                 const { handler, eventName, abi } =
